Initial structure breakout
