+++ lib/gemstone/armaments/weapon_stats_ohe.rb
@@ -1,0 +1,9 @@
+module Lich
+  module Gemstone
+    module Armaments
+      module WeaponStats
+        # stats table
+      end
+    end
+  end
+end
